pyproject.toml
# SPDX-PackageName: mastodon-apps
# SPDX-PackageSupplier: Ryan Finnie <ryan@finnie.org>
# SPDX-PackageDownloadLocation: https://github.com/rfinnie/mastodon-apps
# SPDX-FileCopyrightText: © 2023 Ryan Finnie <ryan@finnie.org>
# SPDX-License-Identifier: MPL-2.0

[project]
name = "mastodon-apps"
authors = [
    {name = "Ryan Finnie", email = "ryan@finnie.org"},
]
description = "Mastodon apps"
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MPL-2.0"}
dependencies = [
    "Pillow",
    "PyYAML",
    "Unidecode",
    'backports.zoneinfo; python_version<"3.9"',
    "beautifulsoup4",
    "blurhash-python",
    "lxml",
    "python-dateutil",
    "requests",
    "urlextract",
]
dynamic = ["version"]

[tool.setuptools.packages.find]
include = [
    "mastodon_apps",
    "mastodon_apps.mastodon",
]

[project.scripts]
mastodon-app-eightfortyseven = "mastodon_apps.eightfortyseven:main"
mastodon-app-fish = "mastodon_apps.fish:main"
mastodon-app-jucika = "mastodon_apps.jucika:main"
mastodon-app-printer = "mastodon_apps.printer:main"

[tool.black]
line-length = 132
